Revert "Add top banner to trial projects with link to schedule tour"
