+++ corehq/apps/analytics/templates/analytics/initial/hubspot.html
@@ -1,3 +1,2 @@
 {% load hq_shared_tags %}
 {% initial_analytics_data 'hubspot.apiId' ANALYTICS_IDS.HUBSPOT_API_ID %}
-{% initial_analytics_data 'hubspot.isDemoVisible' is_demo_visible %}
