+++ corehq/apps/hqwebapp/static/hqwebapp/less/_hq/forms.less
@@ -580,16 +580,20 @@ table .ko-inline-edit .read-only:hover {
 }
 
 .form-bubble-purple {
-  background-color: @color-purple-dark;
-  color: @color-purple-dark-inverse;
+  // temporary colors while palette changes are worked out
+  @purple: #43467F;
+  @off-white: #E3D0FF;
+
+  background-color: @purple;
+  color: @off-white;
 
   a.btn-purple {
-    background-color: @color-purple-dark;
+    background-color: @purple;
     color: @component-active-color;
-    border: 1px solid @color-purple-dark-inverse;
+    border: 1px solid @off-white;
 
     &:hover {
-      background-color: lighten(@color-purple-dark, 5);
+      background-color: lighten(@purple, 5);
     }
   }
 }
