+++ corehq/apps/hqwebapp/static/hqwebapp/less/_hq/includes/variables.less
@@ -85,12 +85,3 @@
 
 @input-border-radius-large:       5px;
 @cursor-disabled:                 'not-allowed';
-
-
-// Expanded color palette
-
-// this is the purple color from the demo button at signup.
-// todo standardize color names a little better. expand palette
-@color-purple-dark: #43467F;
-@color-purple-dark-inverse: #E3D0FF;
-
