+++ corehq/apps/hqwebapp/static/hqwebapp/less/style-imports.less
@@ -21,7 +21,6 @@
 @import "_hq/navbar.less";
 @import "_hq/navs.less";
 @import "_hq/notifications.less";
-@import "_hq/optimize.less";
 @import "_hq/pagination.less";
 @import "_hq/panels.less";
 @import "_hq/popovers.less";
@@ -33,3 +32,4 @@
 @import "_hq/typography.less";
 @import "_hq/svg.less";
 @import "_hq/utils.less";
+@import "_hq/optimize.less";
