+++ corehq/apps/hqwebapp/templates/hqwebapp/base.html
@@ -1,10 +1,7 @@
 {% load menu_tags %}{% load i18n %}{% load hq_shared_tags %}{% load cache %}{% load compress %}{% load statici18n %}<!DOCTYPE html>
 {% get_current_language as LANGUAGE_CODE %}
 {% requirejs_main %}
-<!--[if lt IE 7]><html lang="{{ LANGUAGE_CODE }}" class="lt-ie9 lt-ie8 lt-ie7"><![endif]-->
-<!--[if IE 7]><html lang="{{ LANGUAGE_CODE }}" class="lt-ie9 lt-ie8"><![endif]-->
-<!--[if IE 8]><html lang="{{ LANGUAGE_CODE }}" class="lt-ie9"><![endif]-->
-<!--[if gt IE 8]><!--><html lang="{{ LANGUAGE_CODE }}"><!--<![endif]-->
+<html lang="{{ LANGUAGE_CODE }}">
     <head>
         {% if ANALYTICS_IDS.GTM_ID %}
             {# This is fine as an inline script; it's independent of all HQ code and all third-party libraries #}
@@ -182,9 +179,6 @@
         {% endif %}
         {% block background_content %}{% endblock %}
         {% maintenance_alert %}
-        {% if request.user.is_authenticated and is_demo_visible %}
-          {% include "hqwebapp/partials/get_demo_banner.html" %}
-        {% endif %}
         <!-- .hide used for A/B test of account creation page -->
         <div class="hq-container{% if is_register_user %} molly{% endif %}">
             {% block navigation %}
@@ -508,8 +502,65 @@
           <script src="{% static 'hqwebapp/js/validators.ko.js' %}"></script>
         {% endif %}
 
-        {% if is_demo_visible %}
-          {% include "hqwebapp/partials/get_demo_modals.html" %}
+        {% if not request.user.is_authenticated and is_saas_environment and ANALYTICS_IDS.HUBSPOT_API_ID %}
+
+          <div class="modal fade" id="cta-form-get-demo">
+            <div class="modal-dialog">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">{% trans 'Close' %}</span></button>
+                  <h4 class="modal-title">{% trans "Get a Demo of CommCare" %}</h4>
+                </div>
+                <div class="modal-body form-horizontal">
+                  <div class="form form-horizontal" id="get-demo-cta-form-body">
+
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+
+
+          <div class="modal fade"
+               id="cta-form-get-demo-new">
+            <div class="modal-dialog">
+              <div class="modal-content">
+                <div class="modal-header">
+                  <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">{% trans 'Close' %}</span></button>
+                  <h4 class="modal-title">{% trans "Request a CommCare Demo" %}</h4>
+                </div>
+                <div class="modal-body form-horizontal">
+
+                  {# Hubspot Form #}
+                  <div class="form form-horizontal"
+                           id="get-demo-cta-form-content"></div>
+
+                  {# Schedule Once Demo #}
+                  <div id="get-demo-cta-calendar-content" style="display: none;">
+                    <p>
+                      {% blocktrans %}
+                        Select a time to see CommCare in action and review
+                        questions with our team.
+                      {% endblocktrans %}
+                    </p>
+                    <div id="SOIDIV_commcaredemoform"
+                         data-so-page="commcaredemoform"
+                         data-style="border: 1px solid #d8d8d8; min-width: 290px; max-width: 900px;"
+                         data-psz="11"></div>
+                  </div>
+
+                </div>
+              </div>
+            </div>
+          </div>
+
+          <!--[if lte IE 8]>
+          <script charset="utf-8" src="//js.hsforms.net/forms/v2-legacy.js"></script>
+          <![endif]-->
+          <script charset="utf-8" src="//js.hsforms.net/forms/v2.js"></script>
+
+          <script src="{% static 'analytix/js/hubspot.get_demo.js' %}"></script>
+
         {% endif %}
 
         {# Knockout component templates #}
