--- corehq/apps/hqwebapp/templates/hqwebapp/partials/get_demo_modals.html
@@ -1,58 +1,0 @@
-{% load i18n %}
-
-{# note: remove js from analytix/hubspot.js once A/B test is complete #}
-<div class="modal fade" id="cta-form-get-demo">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">{% trans 'Close' %}</span></button>
-        <h4 class="modal-title">{% trans "Get a Demo of CommCare" %}</h4>
-      </div>
-      <div class="modal-body form-horizontal">
-        <div class="form form-horizontal" id="get-demo-cta-form-body">
-
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
-
-<div class="modal fade"
-     id="cta-form-get-demo-new">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span><span class="sr-only">{% trans 'Close' %}</span></button>
-        <h4 class="modal-title">
-          {% if not request.user.is_authenticated %}
-            {% trans "Request a CommCare Demo" %}
-          {% else %}
-            {% trans "Get a Tour of CommCare" %}
-          {% endif %}
-        </h4>
-      </div>
-      <div class="modal-body form-horizontal">
-
-        {# Hubspot Form #}
-        <div class="form form-horizontal"
-                 id="get-demo-cta-form-content"></div>
-
-        {# Schedule Once Demo #}
-        <div id="get-demo-cta-calendar-content" style="display: none;">
-          <p>
-            {% blocktrans %}
-              Select a time to see CommCare in action and review
-              questions with our team.
-            {% endblocktrans %}
-          </p>
-          <div id="SOIDIV_commcaredemoform"
-               data-so-page="commcaredemoform"
-               data-style="border: 1px solid #d8d8d8; min-width: 290px; max-width: 900px;"
-               data-psz="11"></div>
-        </div>
-
-      </div>
-    </div>
-  </div>
-</div>
